Created data merger python code and finished the  last version of the montecarlo code
Added the folder from the last session, with the MC3 code which is the final version of the code. The only things missing are to compile for different time intervals and to complete the data merger code to take the data from the different txt files and to write it in a neat and addequate way in a different file.